Added two use cases for useRef.
In the first use case, whenever a user adds a new item to the list, the input field will automatically receive focus.

In the second use case, the character limit of the input box is set to an arbitrary number (in this case, 99), and a character count display is shown.